Organize project directory structure
Finish organizing project directory structure

* Extract common page components from "home.js" to "Buttons.js" and
"Widgets.js" files
* Move all page component files to new "pages" folder
