+++ src/main/App.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import "./App.css";
+
 import Banner from "./components/Banner.js";
-import Home from "./components/Home.js";
-import Deposit from "./components/Deposit.js";
-import SignOut from "./components/SignOut.js";
-import Backup from "./components/Backup.js";
-import Withdraw from "./components/Withdraw.js";
+import Home from "./components/pages/Home.js";
+import Deposit from "./components/pages/Deposit.js";
+import SignOut from "./components/pages/SignOut.js";
+import Backup from "./components/pages/Backup.js";
+import Withdraw from "./components/pages/Withdraw.js";
+
 import { BrowserRouter as Router, Route, NavLink, Switch} from "react-router-dom";
 
 class App extends React.Component {
