+++ src/main/components/buttons.css
@@ -1,0 +1,41 @@
+.link_button {
+  height: 95px;
+  display: block;
+  width: var(--item-width);
+  color: white;
+  position: relative;
+  margin: 0 auto;
+}
+
+.button_text {
+  position: absolute;
+  margin: 0;
+  top:50%;
+  left:50%;
+  transform: translate(-50%, -50%);
+  text-align: center;
+  font-weight: bold;
+}
+
+.blue_button {
+  background-color: #56c1ff;
+  font-size: 34px;
+  border: 4px solid black;
+}
+
+.clear_button {
+  background-color: inherit;
+  border: none;
+  background: none;
+  font-size: 26px;
+  padding-bottom: 16px;
+}
+
+.regenerate_button {
+  right: 0px;
+  font-size: 22px;
+}
+
+.regenerate_button_left_spacer {
+  width: 758px;
+}
