--- src/main/components/home.css
@@ -1,142 +1,0 @@
-:root {
-  --item-width: 860px;
-}
-
-#home {
-  display: grid;
-  grid-template-areas:
-  ". . ."
-  ". page_box .";
-  grid-template-rows: 110px auto;
-  grid-template-columns: 1fr 990px 1fr;
-  padding: 0;
-  margin: 0;
-}
-
-.page_box {
-  grid-row-gap: 0px;
-  grid-area: page_box;
-  width: inherit;
-  display: grid;
-  grid-template-areas:
-  "title"
-  "sub_title"
-  "header"
-  "regenerate_button"
-  "regenerate_button_bottom_margin"
-  "main_content"
-  "save_phrase_box_bottom_margin"
-  "blue_button"
-  "clear_button";
-  background-color: #017b76;
-}
-
-.save_phrase_box_bottom_margin {
-  grid-area: save_phrase_box_bottom_margin;
-  height: 28px;
-}
-.regenerate_button_bottom_margin {
-  grid-area: regenerate_button_bottom_margin;
-  height: 16px;
-}
-
-.page_box > div {
-  margin: 0 auto;
-}
-
-.title {
-  grid-area: title;
-  font-size: 45px;
-  padding-bottom: 36px;
-  padding-top: 40px;
-  margin: 0;
-  line-height: 33px;
-}
-
-.sub_title {
-  grid-area: sub_title;
-  font-weight: bold;
-  font-size: 26px;
-  padding-bottom:72px;
-  margin: 0;
-}
-
-.header {
-  grid-area: header;
-  font-size: 35px;
-  padding-top: 44px;
-  margin: 0;
-  line-height: 33px;
-}
-
-.regenerate_button_container {
-  grid-area: regenerate_button;
-  display: flex;
-  flex-direction: row;
-}
-
-.regenerate_button {
-  right: 0px;
-  font-size: 22px;
-}
-
-.regenerate_button_left_spacer {
-  width: 758px;
-}
-
-.main_content {
-  grid-area: main_content;
-  width: var(--item-width);
-}
-
-.link_button {
-  height: 95px;
-  display: block;
-  width: var(--item-width);
-  color: white;
-  position: relative;
-  margin: 0 auto;
-}
-
-.button_text {
-  position: absolute;
-  margin: 0;
-  top:50%;
-  left:50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-  font-weight: bold;
-}
-
-.blue_button {
-  grid-area: blue_button;
-  background-color: #56c1ff;
-  font-size: 34px;
-  border: 4px solid black;
-}
-
-.clear_button {
-  grid-area: clear_button;
-  background-color: inherit;
-  border: none;
-  background: none;
-  font-size: 26px;
-  padding-bottom: 16px;
-}
-
-.sub_page_link_button {
-  height: 72px;
-}
-
-.save_phrase_box {
-  background-color: #d6d5d5;
-  font-size: 31px;
-  color: black;
-  line-height: 41px;
-  border: 4px solid black;
-}
-
-.save_phrase_box>div {
-  padding-top: 20px;
-  padding-bottom: 62px;
-}
