+++ src/main/components/pages/SignOut.js
