+++ src/main/components/widgets.css
@@ -1,0 +1,54 @@
+.page_box > div {
+  margin: 0 auto;
+}
+
+.save_phrase_box_bottom_margin {
+  height: 28px;
+}
+.regenerate_button_bottom_margin {
+  height: 16px;
+}
+
+.title {
+  font-size: 45px;
+  padding-bottom: 36px;
+  padding-top: 40px;
+  margin: 0;
+  line-height: 33px;
+}
+
+.sub_title {
+  font-weight: bold;
+  font-size: 26px;
+  padding-bottom:72px;
+  margin: 0;
+}
+
+.header {
+  font-size: 35px;
+  padding-top: 44px;
+  margin: 0;
+  line-height: 33px;
+}
+
+.regenerate_button_container {
+  display: flex;
+  flex-direction: row;
+}
+
+.main_content {
+  width: var(--item-width);
+}
+
+.save_phrase_box {
+  background-color: #d6d5d5;
+  font-size: 31px;
+  color: black;
+  line-height: 41px;
+  border: 4px solid black;
+}
+
+.save_phrase_box>div {
+  padding-top: 20px;
+  padding-bottom: 62px;
+}
